requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
apscheduler==3.10.4
python-telegram-bot==21.7
playwright==1.49.0
anthropic==0.40.0
python-multipart==0.0.17
